Fix: git gets lost when we try to checkout a duplicate branch.
Suppose we want to clone the repository http://foo.bar/repo and checkout
the bug-XXX branch:

    tc-vcs checkout repo http://foo.bar/repo http://foo.bar/repo bug-XXX

This will fail because there will be the refs/remotes/origin/bug-XXX
and the refs/remotes/tc-vcs-remote/bug-XXX refs. "git checkout bug-XXX"
doesn't know which one to checkout.

We now play safe and check if the ref exists locally and use its
remote if it exists, otherwise use tc-vcs-remote.