Reduce number of retries
If we can't clone, chances are the network or server is having a bad
time. For the latter, several retries could effectively flood a
server with excess traffic, further bringing the struggling server to
its knees.

This commit recognizes that Taskcluster constitutes the thundering herd
and reduces the retry count on clones to simply 1. If the server is
having problems, jobs will fail fast instead of carrying out a prolonged
barrage of clone attempts.

While we were here, we reduced Mercurial's checkoutRevision retry count
as well. This command should be idempotent, so extra retries are likely
worthless.