Pull remote when bootstrapped from a cache
Before this commit, if a clone was serviced from the artifact cache, the
local clone would potentially be very out of date with the remote
(depending on when the artifact cache was populated). If the base/clone
URL was different from the repo URL, the incremental difference would be
fetched from the repo URL, not the base URL. In the ideal world, this
"just works." However, for Mozilla's "try" repository, this is not the
case. This repository has tens of thousands of heads and its repository
storage is structured in such a way that fetching hundreds or thousands
of changesets is slow. It is therefore 2-3x more expensive to pull from
this repository than from the base repository.

This commit will almost certainly result in an increase in total
repository traffic. However, that increase will be directed at the base
repository, not the final repository. This is acceptable because the
base repository is often provisioned to scale better than a one-off
fork. It will also move load away from the Try repository, which is
highly desired.