Reduce the amount of parallelism when calling repo sync
When calling repo sync taskcluster-vcs will try to do things in the
most expedient ways, which includes running it with -j100 by default
and also calling repo sync for each of the projects in parallel.

This is believed to cause numerous connections to the git servers.  When
caches are used, and only changes need to be synced, calling `repo sync
-j1` on the parent project as a whole does not seem to be slower, and in
many trials faster.

For those that would still like to speed up their syncs, an option can
now be passed in to override the default value of 1.